base browse page refined as server side & client's
