+++ bigbytes/src/app/browse/page.tsx
@@ -1,24 +1,8 @@
-"use client";
+import { getAllCMSFileSlugs } from "@/lib/cms";
+import Results from "./results";
 
-import { useSelector } from "react-redux";
-import type { RootState } from "@/redux/store/store";
-import Development from "@/app/under-dev";
-import clsx from "clsx";
-
-type ResultProps = {
-  classList?: {
-    tw?: string[];
-    cs?: string[];
-  }
-}
-
-export default function Result({ classList }: ResultProps) {
-  const cls = clsx(classList?.tw, classList?.cs);
-  const underDevelopment = useSelector((state: RootState) => state.development?.pages?.browse);
-    
-  if(underDevelopment) return <Development/>
-
-  return <main id="Result" className={cls}>
-    <h2>Result Page</h2>    
-  </main>
+export default async function Page() {
+  const slugs = await getAllCMSFileSlugs();
+  
+  return <Results slugList={slugs} />
 }
